+++ tests/purgeUnusedStyles.test.js
@@ -105,7 +105,7 @@ test('purges unused classes', () => {
   )
 })
 
-test('purge patterns are resolved relative to the config file', () => {
+test('purge patterns are resolved relative to the current working directory', () => {
   return inProduction(
     suppressConsoleLogs(() => {
       const inputPath = path.resolve(`${__dirname}/fixtures/tailwind-input.css`)
@@ -567,6 +567,32 @@ test(
           options: {
             content: [path.resolve(`${__dirname}/fixtures/**/*.html`)],
             safelist: ['md:bg-green-500'],
+          },
+        },
+      }),
+    ])
+      .process(input, { from: withTestName(inputPath) })
+      .then((result) => {
+        expect(result.css).toContain('.md\\:bg-green-500')
+        assertPurged(result)
+      })
+  })
+)
+
+test(
+  'proxying purge.safelist to purge.options.safelist works',
+  suppressConsoleLogs(() => {
+    const inputPath = path.resolve(`${__dirname}/fixtures/tailwind-input.css`)
+    const input = fs.readFileSync(inputPath, 'utf8')
+
+    return postcss([
+      tailwind({
+        ...config,
+        purge: {
+          enabled: true,
+          safelist: ['md:bg-green-500'],
+          options: {
+            content: [path.resolve(`${__dirname}/fixtures/**/*.html`)],
           },
         },
       }),